fix: optimize transaction history
