+++ package.json
@@ -35,6 +35,7 @@
         "react-i18next": "^15.1.3",
         "react-icons": "^5.4.0",
         "react-markdown": "^9.0.1",
+        "react-window": "^1.8.11",
         "socket.io-client": "^4.8.1",
         "styled-components": "^6.1.12",
         "tauri-plugin-sentry-api": "^0.2.0",
@@ -53,6 +54,7 @@
         "@types/node": "^22.10.1",
         "@types/react": "^18.3.14",
         "@types/react-dom": "^18.3.2",
+        "@types/react-window": "^1.8.8",
         "@types/uuid": "^10.0.0",
         "@typescript-eslint/parser": "^8.3.0",
         "@vitejs/plugin-react": "^4.3.4",
