+++ src/containers/main/SideBar/components/Wallet/History.tsx
@@ -1,10 +1,12 @@
 import { useTranslation } from 'react-i18next';
 import { useWalletStore } from '@app/store/useWalletStore';
 import { CircularProgress } from '@app/components/elements/CircularProgress';
+import { FixedSizeList as List } from 'react-window';
 
 import { ListLabel } from './HistoryItem.styles';
 import { HistoryContainer, HistoryPadding } from './Wallet.styles';
 import HistoryItem from './HistoryItem';
+import { memo } from 'react';
 
 const container = {
     hidden: { opacity: 0, height: 0 },
@@ -14,10 +16,16 @@ const container = {
     },
 };
 
-export default function History() {
+const History = () => {
     const { t } = useTranslation('sidebar', { useSuspense: false });
     const isTransactionLoading = useWalletStore((s) => s.isTransactionLoading);
     const transactions = useWalletStore((s) => s.transactions);
+
+    const Row = ({ index, style }) => (
+        <div style={style}>
+            <HistoryItem key={transactions[index].tx_id} item={transactions[index]} />
+        </div>
+    );
 
     return (
         <HistoryContainer initial="hidden" animate="visible" exit="hidden" variants={container}>
@@ -26,9 +34,13 @@ export default function History() {
                 {isTransactionLoading && !transactions?.length ? (
                     <CircularProgress />
                 ) : (
-                    transactions.map((tx) => <HistoryItem key={tx.tx_id} item={tx} />)
+                    <List height={310} itemCount={transactions.length} itemSize={62} width="100%">
+                        {Row}
+                    </List>
                 )}
             </HistoryPadding>
         </HistoryContainer>
     );
-}
+};
+
+export default memo(History);
