+++ src/containers/main/SideBar/components/Wallet/Wallet.tsx
@@ -1,12 +1,7 @@
-import { useCallback, useState, useEffect } from 'react';
+import { useCallback, useState } from 'react';
 import { useBlockchainVisualisationStore } from '@app/store/useBlockchainVisualisationStore';
 import { useTranslation } from 'react-i18next';
 import { AnimatePresence } from 'framer-motion';
-import { IoEyeOutline, IoEyeOffOutline } from 'react-icons/io5';
-import CharSpinner from '@app/components/CharSpinner/CharSpinner.tsx';
-import { formatNumber, FormatPreset } from '@app/utils/formatters.ts';
-import { Typography } from '@app/components/elements/Typography.tsx';
-import { Stack } from '@app/components/elements/Stack.tsx';
 
 import { useWalletStore } from '@app/store/useWalletStore.ts';
 import { usePaperWalletStore } from '@app/store/usePaperWalletStore.ts';
@@ -17,17 +12,14 @@ import SyncTooltip from './SyncTooltip/SyncTooltip.tsx';
 import History from './History.tsx';
 
 import {
-    BalanceVisibilityButton,
     CornerButton,
     CornerButtonBadge,
     ScrollMask,
     SidebarCover,
-    WalletBalance,
-    WalletBalanceContainer,
-    WalletBalanceWrapper,
     WalletContainer,
     WalletCornerButtons,
 } from './Wallet.styles.ts';
+import WalletBalanceMarkup from './WalletBalanceMarkup.tsx';
 
 export default function Wallet() {
     const { t } = useTranslation('sidebar', { useSuspense: false });
@@ -41,26 +33,9 @@ export default function Wallet() {
     const recapCount = useBlockchainVisualisationStore((s) => s.recapCount);
     const setRecapCount = useBlockchainVisualisationStore((s) => s.setRecapCount);
 
-    const [showBalance, setShowBalance] = useState(true);
     const [showHistory, setShowHistory] = useState(false);
-    const [showLongBalance, setShowLongBalance] = useState(false);
-    const [animateNumbers, setShowAnimateNumbers] = useState(true);
 
     const fetchTx = useFetchTx();
-    const formatted = formatNumber(balance || 0, FormatPreset.TXTM_COMPACT);
-    const formattedLong = formatNumber(balance || 0, FormatPreset.TXTM_LONG);
-
-    const sizingLong = useCallback(() => {
-        const baseSize = 50;
-        const step = 1.75;
-        const maxLength = 20;
-        const length = Math.min(formattedLong.length, maxLength);
-        if (length <= 9) return baseSize;
-        return baseSize - (length - 6) * step;
-    }, [formattedLong.length]);
-
-    const toggleBalanceVisibility = () => setShowBalance((prev) => !prev);
-    const displayValue = balance === null ? '-' : showBalance ? formatted : '*****';
 
     const handleShowClick = useCallback(async () => {
         if (balance && !transactions.length && !isTransactionLoading) {
@@ -76,76 +51,6 @@ export default function Wallet() {
     const handleSyncButtonClick = () => {
         setShowPaperWalletModal(true);
     };
-
-    const handleMouseOver = () => {
-        setShowAnimateNumbers(false);
-        setShowLongBalance(true);
-    };
-
-    const handleMouseOut = () => {
-        setShowAnimateNumbers(false);
-        setShowLongBalance(false);
-    };
-
-    useEffect(() => {
-        if (!animateNumbers) {
-            const timer = setTimeout(() => {
-                setShowAnimateNumbers(true);
-            }, 300);
-
-            return () => clearTimeout(timer);
-        }
-    }, [animateNumbers]);
-
-    const balanceMarkup = (
-        <WalletBalanceContainer>
-            <Stack direction="row" alignItems="center">
-                <Typography variant="span" style={{ fontSize: '15px' }}>
-                    {t('wallet-balance')}
-                </Typography>
-                <BalanceVisibilityButton onClick={toggleBalanceVisibility}>
-                    {showBalance ? (
-                        <IoEyeOffOutline size={14} color="white" />
-                    ) : (
-                        <IoEyeOutline size={14} color="white" />
-                    )}
-                </BalanceVisibilityButton>
-            </Stack>
-            <WalletBalanceWrapper onMouseOver={handleMouseOver} onMouseOut={handleMouseOut}>
-                <AnimatePresence mode="popLayout">
-                    {!showLongBalance || !showBalance ? (
-                        <WalletBalance
-                            initial={{ opacity: 0 }}
-                            animate={{ opacity: 1 }}
-                            exit={{ opacity: 0 }}
-                            key="compressed-number"
-                        >
-                            <CharSpinner
-                                value={displayValue}
-                                variant="simple"
-                                fontSize={50}
-                                animateNumbers={animateNumbers}
-                            />
-                        </WalletBalance>
-                    ) : (
-                        <WalletBalance
-                            initial={{ opacity: 0 }}
-                            animate={{ opacity: 1 }}
-                            exit={{ opacity: 0 }}
-                            key="full-number"
-                        >
-                            <CharSpinner
-                                value={formattedLong}
-                                variant="simple"
-                                fontSize={sizingLong()}
-                                animateNumbers={animateNumbers}
-                            />
-                        </WalletBalance>
-                    )}
-                </AnimatePresence>
-            </WalletBalanceWrapper>
-        </WalletBalanceContainer>
-    );
 
     const showCount = Boolean(recapCount && recapCount > 0 && !showHistory);
 
@@ -174,7 +79,7 @@ export default function Wallet() {
                     ) : null}
                 </WalletCornerButtons>
 
-                {balanceMarkup}
+                <WalletBalanceMarkup />
 
                 <AnimatePresence mode="wait">
                     {showHistory ? (
